Update main.py
One way to make this code more efficient would be to only run the alembic upgrade on startup if the database is not already up to date. To do this, you could check the current alembic version before running the upgrade, and only run the upgrade if the current version is not the latest version.

Here's an example of how you could modify the alembic_upgrade function to do this:
@app.on_event("startup")
def alembic_upgrade():
    logger.info("Checking alembic version on startup")
    alembic_ini_path = os.path.join(os.path.dirname(__file__), "alembic.ini")
    alembic_cfg = config.Config(alembic_ini_path)
    alembic_cfg.set_main_option("sqlalchemy.url", settings.DATABASE_URI)

    # Check current alembic version
    from sqlalchemy import create_engine
    from alembic.runtime.environment import EnvironmentContext
    engine = create_engine(settings.DATABASE_URI)
    with engine.connect() as connection:
        context = EnvironmentContext(alembic_cfg, connection)
        current_version = context.get_current_revision()

    # Upgrade alembic if current version is not the latest version
    if current_version != command.get_current_head(alembic_cfg):
        logger.info("Upgrading alembic on startup")
        command.upgrade(alembic_cfg, "head")
        logger.info("Successfully upgraded alembic on startup")
    else:
        logger.info("Alembic already up to date")
This code first checks the current alembic version by creating a connection to the database and using the EnvironmentContext class from the alembic library. If the current version is not the latest version, the code runs the alembic upgrade. Otherwise, it simply logs a message to indicate that the