website: show language in text area
When the language hasn't been detected (too little text) the langauge that it will be submitted as is shown. When we detect the language is not what will be submitted we show the detected language in red. The tooltip informs the user that they don't appear to be writing in the correct language.

Note: this replaces the modal dialog that we use to popup if the language appeared to differ.