View chat list (#1915)
Close #1881

Barebones chat list UI.

The chats we have only contain the ID, we need to expand them to show
more stuff so that the users can recongnise which chat it is. Maybe
showing the last message and the date of the message would be helpful.

I tried as much as possible to avoid potential conflicts with #1913. 


Known issues: Currently we don't fetch all messages for a given chat ID
& therefore we cannot continue a previous chat (because we need the id
of the last message), this could be fixed easily by adding a fetch call
to `ChatConversion` or whichever equivalent component.



![image](https://user-images.githubusercontent.com/24505302/221964414-42a904bd-7ae7-4448-99e3-61e2892063d0.png)


![image](https://user-images.githubusercontent.com/24505302/221965689-3a38d7a6-6a33-4353-8a07-91e9f5ce73c4.png)