Updated team page to fit visually more the the rest of the pages, added a button to the call-to-action section which links to the team page. (#2126)
Hey, as discussed here
https://github.com/LAION-AI/Open-Assistant/issues/1703 ...
1. I made a few cosmetical changes to the new team page (it looks now
similar to the ToS and Privacy-Policy page):

![image](https://user-images.githubusercontent.com/25230234/226197002-a50887bf-a1e5-41f6-9202-f325e21f067e.png)
I'm not sure why the ``team_message`` translation doesn't work but it
was like this on the main branch already.
The only "issue" with this design is that you can see the boxes of the
team members streching. But I still find it visually nicer as before.

2. I also added a button next to the Discord and GitHub buttons in the
Call-To-Action section which links to the team page so it that it's more
accessible. Let me know what you think about that:

![image](https://user-images.githubusercontent.com/25230234/226196938-9e6ff60a-84a8-4646-a5d8-97f2efd9bc51.png)

---------

Co-authored-by: Yannic Kilcher <yk@users.noreply.github.com>
Co-authored-by: theopfr <teddypeifer@gmailc.com>
Co-authored-by: notmd <tinhmeo10@gmail.com>