Add support for Cerebras-GPT for training (#2276)
This adds configs for the 13B and 6.7B Cerebras models but with the
necessary code changes made it should be easy enough to add configs for
the smaller models too if desired. The tokenizer seems to be the GPT-2
fast tokenizer from HuggingFace so the special tokens have been
configured for that.