Implement auto scroll when stream message (#2839)
After trying a bunch of methods I found this is the easiest way to do
it. Other approaches are way more complicated.
Also found this lib https://github.com/compulim/react-scroll-to-bottom
OAI also uses it, but it doesn't support SSR and unmaintained for years.
This solution is not really reliable, if you paste a prompt with many
lines (4 or above), it doesn't work. It would be if someone can patch
the `react-scroll-to-bottom` to support SSR.