Add Dataset with Russian Riddles (#3074)
Dataset Card for Russian riddles with answers with 377 entries.

### Dataset Summary

Contains parquet of QnA with riddle & answer pairs.

Each row consists of

* INSTRUCTION
* RESPONSE
* SOURCE
* METADATA (json with language).

### Licensing Information

Data is scrapped from several sites. Since most of the riddles and
answers are publicly available and popular, any ToS and licensing of the
sites themselves is irrelevant. I reserve the right to put a public and
permissive license.
Moreover, there was no licensing information on these sites, which makes
sense, due to the public availability and prominence of the content they
provide.

### Acknowledgements

Thanks Freddie#5762 for providing this data!
He mentioned these URLs:
- https://azbyka.ru/deti/logicheskie-i-zanimatelnye-zadachi
- https://bbf.ru/riddles/

---

Resolves https://github.com/LAION-AI/Open-Assistant/issues/3078