Add Bavarian ("Bairisch") Language (#3033)
Why add it:
1. According to [ISO
639-3](https://de.wiktionary.org/wiki/Hilfe:Sprachk%C3%BCrzel) it is
recognized as its own Language as {bar}.
2. The Linguistic Distance between
Swedish-Norwegish-Danish;Serbian-Croatian;Czech-Slovakian is greater
than the Difference between Bavarian and Standardgerman. (Standardgerman
is supported in OA as "German", but **Bavarian is not a Dialect of
Standardgerman**!) Therefore it can not be a simple Dialect because the
Linguistic Distance is greater and [Bavarian is older than
Standardgerman!](https://bar.wikipedia.org/wiki/Oidboarisch).
3. In South Tyrol, Austria and Switzerland Dialects of Bavarian and
Allemanic are spoken since the Migration Period and South Tyrolean
German, Austrian German and Swiss German officially used in School and
any written Form. These are Languages Variations of German which are
EQUAL to the Standardgerman of Germany, have their own Rules and Words.
I list this here so People know that "German" written and spoken in
Germany is not the only correct way to spell and write a
Standardgermanvariation. But i do not care about these 4 Standards,
because i consider them to be foreign Languages to me.
4. There exist already many Websites, Books and Text which investigate
and document Bavarian and its Dialects, namely:
- the bavarian dictionary (https://bwb.badw.de/das-projekt.html)
(BAYRISH WORDS FROM BAYERN! but a good start)
- a list of 3300 southtyrolean (or better southbavarian/bavarian) Words
by @brgr (https://gist.github.com/brgr/8fed41cadab07d25383d5c909d2e177d)
(i looked over them and know most, they are not just used in ST but also
North- and EastTyrol, Salzburg and Carinthia, the rest of Austria and
Bavaria aswell atleast some)
- the whole bavarian wikipedia
(https://bar.wikipedia.org/wiki/Wikipedia:Hoamseitn)
- the southtyrolean Lexikon (https://suedtiroler-lexikon.org/) with 1700
Words, mostly congruent with the github by brgr and Oschpele
- (https://www.ritten.org/oschpele/)
5. Bavarian is recognized by
[Glottolog](https://glottolog.org/resource/languoid/id/bava1246),
[Ethnologue](https://www.ethnologue.com/language/bar/) and as an
Endangered Language. (thats why im here :) )
6. IT HAS RULES that are already described in great Detail over 100
Books, 60 of them are listed on Glottolog here:
https://glottolog.org/resource/languoid/id/bava1246


The more Interdisciplinary and everyday Stuff:
1. In Southtyrol when we write a SMS or Whatsapp, a simple Note or speak
on the Phone we do never ever talk or write in "German", we do use
exclusivly a Dialect of Bavarian or to make it simple Bavarian. I think
this applies to the other Tyrol, Austria and Bavaria aswell.
2. This brings me to my next Issue: Today "German" means in
everydayspeak by the average Person "Standardgerman". The German
Tourists talk German it is said, which to be correctly is now equivalent
to "German". **So if German today means Standardgerman and Bavarian is
older than Standardgerman than Bavarian can not be shoved under the
German Tab or be a Dialect of German=Standardgerman.** The German
Tourists are called the Germans, Swiss the Swiss and Austrians the
Austrians; therefore no Germafeeling exists outside Germany anymore
supported by Ernst Bruckmüller (Austrian Selfidentification) and
(https://www.provinz.bz.it/familie-soziales-gemeinschaft/integration/images/899470_sprachbarometer_2014.pdf)
So where no more Germans are, they can not be German. Identity=Language.
Every European Country speaks a Language that is called after the
Countrys Name, but the "German Lands" (obsolete since 1530 1870, 1919
and 1945 depends on Definition) speak all German? This is not how it
works. ( According to official studies by ASTAT in "Sprachbarometer
2014" Southtyroleans were asked how they feel: 80% of "Natives"
(non-Ladinians or Italians) feel South Tyrolean, 4% austrian and
4%german.)
3. The vast majority of germans who speak standardgerman do not
understand my written language, speech or words. Only the Bavarians,
Swiss and Austrians do but they speak Bairisch and Swiss.
4. Putting South Tyrolean or Bavarian under German instead of its own
Language is a big Identity Issue for the South Tyrolean Natives,
Austrians and "Swissspeaking Swiss" because they today live in a Romance
speaking State, italy, as a minority; do not share the same culture and
mentality as the Germans (South tyroleans are much closer to the
bairisch speaking Bavarians and Austrians and Allemannic speaking Swiss
and Rhaetoromance speakers like the Ladins and eqiivalent in
Switzerland:Many words and place Names and speech patterns still exist
from Latin times in Zillertal and Vinschgau; and in Austria even slavic
Familynames, Placenames, Everydaywords...ladin equivalents in
Southtyrol, Eastswitzerland and Tyrol!)
5. (I dont mean this to seriously, these Times happened in the Hitler
Days and before; much less today but still a small Issue.) Calling
Bavarian a Dialect of Standardgerman is an (unconscious) imperialistic
instrument to claim ownership of Lands and Language Authority by the
Germans. (Not my Words but Dr. Stefan Dollinger)
6. South Tyrolean or all the other Bavarian Dialect and Bavarian itself
were never called "southtyrolean Deutsch" or "bairischdeutsch". See
Dutch, its just Deutsch with a missing Letter but is its own Language
with Dialects.
How can it then be that Dutch is not under Deutsch/German, even if its
clearly the Same Word with a missing Letter?
7.The Backlabelling of putting German/deutsch on the Kaiser of Germany
("Ludwig der Deutsche") and its Population Groups like the
Baiiuvari/Bavarians is not correct. They never saw themselves as German
or that a alloverarching Germaness is over their People or Land. Also
Germanic People were never one Ethnic Group.
8.How can Austrians, Swiss and South Tyroleans be Germans and speak
German (Standardgerman) when they live in different Nations, Cultures,
Mentalites and have different Genes?
9.A German now means only and exclusivly a Citizen of the Federal
Republic of Germany. There is no German Culture or Language ouside of
it. (Expect maybe the Hutterites see themselves still as German but in
the Sense of 1800!)
10.Dutch was by a Prussian Linguist put under the German Branch in 1905,
today no one would dare this and the Hate would be immense. It is just
the same Thing today with Bavarian and its Dialects,

As a Minority Person and Speaker (South Tyrolean in Italy, Bavarian
Language Speaker which is endangered both nationally and
internationally.) this is obviously a very important Issue, ;)