add readme to plugins in oasst (#3273)
could be a nice standard to add a README.md next to each plugin that
lives in the repo. just putting this in as a sort of example. idea being
docs would link to this folder for example and so having readme there
would be useful for more details and stuff that might not be obvious
from the spec json itself.