Moderator Message Editing (#3280)
closes #1005 

Backend changes:
- Added an endpoint to edit messages
- When a message is edited a new entry for the table 'message_revision'
is created which contains the current contents, the user who made that
version, the time that version was created, and the version which came
before that one. The actual message in the messages table will then be
updated to contain the new content, date, user who made the most recent
edit, and a reference to the last version. (This has been changed
slightly, see comment below)

Frontend changes:
- Added a page which shows the context surrounding a specific message
(ie. all messages before and after it, but not messages in different
threads)
- That page has an entry box where the new content is submitted.
- That page can be accessed through the menu of a message or through the
admin message dashboard.