Initial inference documentation pass (#3330)
In the recent meeting it was raised that the codebase is not documented
and hard to understand as a result. This adds some initial
documentation. For now it is quite basic but hopefully will allow a bit
more understanding of some parts of the code "at a glance", reducing the
friction somewhat.