Fix out of bound index in chat paging (#3480)
closes #3478 

`useIsomorphicLayoutEffect` wasn't being called until the first render
meaning if the index could be out of bounds if the parent message was
changed to a tree with less siblings than the previous. A simple min
against the number of siblings prevents the out of bound index and then
`useIsomorphicLayoutEffect ` is called the next frame, setting the
correct thread.