3517 - Send reports to discord via celery task
In debeb353cc0104a0a330637d078b9b0bc41afe84 we reverted the original implementation of the task as it broke the production environment. The problem was that celery tasks need objects that could be JSON serializable and we were passing a `Message` instance, so just passing the details fixed the issue.