Add paper acknowledgement form (#3637)
I know all of this is suboptimal, but I'd like to get people to sign up
to be mentioned asap. Feel free to improve (or to tell me how to
improve) now or in subsequent PRs. I anticipate this form will be going
away in the future, so I'm not sure a lot of work will be really worth
it.

---------

Co-authored-by: AbdBarho <ka70911@gmail.com>