To fix issue 3648 Check whether the size of the embedding layer has been resized, rather than checking if the length of the tokenizer is equal to the size of the embedding layer.
This is because the embedding layer of the model may be resized and padded to a multiple of 16,
but the tokenizer may not be padded, which will cause the freezing layer to fail. However,
at this time, the embedding layer will not be resized and freezing can still be performed.