Update docs for current project status. (#3730)
This PR updates the FAQ, which is currently very out of date, to the
current status of the project. It also adds a notice to the introduction
page in the docs that the project has concluded. Feel free to make any
changes if it isn't worded how you would like.